requirements.txt
# jupyter
pandas
numpy
requests
tiktoken
matplotlib
# langchain
openai
# python-docx
# python-pptx
PyPDF2
# markdown2
# langchain-community
# langchain-core
# sentence-transformers
# faiss-cpu
# langchain-openai
python-dotenv